requirements.txt
opencv-python
pytesseract
numpy
scikit-image
matplotlib
pyqtgraph
pyqt5
pillow
ffmpeg
path.py